bringing modifications from SVN trunk to GIT
Former-commit-id: 3fedb7cb1d263f2fda84ef4871046a7c2eb887cd

+++ TileCalorimeter/TileRec/TileRec/TileHitToNtuple.h
@@ -54,7 +54,7 @@ private:
     int  m_maxLength;
 
     NTuple::Item<int> m_nchan;
-    NTuple::Item<float> m_tolE;
+    NTuple::Item<double> m_tolE;
 
     NTuple::Array<float> m_energy;
     NTuple::Array<float> m_time;
